requirements.txt
matplotlib==3.10.8
numpy==2.3.5
pandas==3.0.1
pillow==12.0.0
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
